create taxonomy group and subcommands
the Taxonomy group has one command for now:
ilab taxonomy diff

ilab taxonomy download will be a future command

create this command, the parent group, and remove the command in lab.py
