+++ .gitignore
@@ -3,6 +3,7 @@ models/
 !src/instructlab/train/lora_mlx/models
 # We recommend using taxonomy here, but not a submodule. Just a work dir.
 taxonomy/
+!src/instructlab/taxonomy/
 # We recommend putting generated files here, but PLEASE don't commit them.
 generated/
 # We recommend outputting training results on Linux here, but PLEASE don't commit them.
